endpoint delete user

+++ src/app/UseCase/User/UserDeleteUseCase.php
@@ -3,7 +3,7 @@
  * @ Author: Tommyprmbd
  * @ Create Time: 2024-06-01 14:11:48
  * @ Modified by: Tommyprmbd
- * @ Modified time: 2024-06-01 14:16:29
+ * @ Modified time: 2024-06-02 01:19:15
  * @ Description:
  */
 
@@ -11,6 +11,7 @@ namespace App\UseCase\User;
 
 use App\Domain\Entity\User;
 use App\Domain\Repository\UserRepositoryInterface;
+use App\Infrastructure\Exception\NotFoundException;
 
 class UserDeleteUseCase
 {
@@ -21,6 +22,12 @@ class UserDeleteUseCase
     }
 
     public function handle(int $id) {
+        // get user by id
+        $user = $this->userRepositoryInterface->findById($id);
+        if (!$user) {
+            throw new NotFoundException("User ID not found.");
+        }
+        
         return $this->userRepositoryInterface->delete($id);
     }
 }
